enap_backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-environ==0.11.0
Pillow==10.2.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-filter==23.5
whitenoise==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.0
django-storages==1.14.2
django-rest-auth==0.9.5
django-allauth==0.61.1
celery==5.3.6
redis==5.0.1
django-cleanup==8.0.0
Faker==22.0.0 